# backend/api.yaml
openapi: "3.0.0"
info:
  title: TeaTimeSystem API Overview
  version: 1.0.0
paths:
  /orders:
    post:
      tags:
        - "發起人-新增訂單"
      summary: "新增訂單"
      description: ""
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/orders"        
      responses:
        '201':
          description: create a new order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orders"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/responses/BadRequest"
        '401':
          description: Unauthoried
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Unauthoried"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NotFound"    
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Internal Server Error"

            
  /restaurants:
    get:
      tags:
        - "發起人-回傳餐廳陣列從資料庫"
      description: ""
      responses:
        '200':
          description: get the restaurant array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/restaurants"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/responses/BadRequest"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NotFound"    
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Internal Server Error"
        # 補完responses

  /orders/search/{invitation_code}:
    get:
      tags:
        - "發起人/參與人-取得指定編號的全部訂單資訊"
      parameters:
        - in: path
          name: invitation_code
          schema:
            type: string
          required: true
      responses:
        '200':
          description: get the infomation of the order with the invitation code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orders"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/responses/BadRequest"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NotFound"    
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Internal Server Error"

  /order/search/{invitation_code}/{participant_id}:
    get:
      tags:
        - "發起人/參與人-操作品項於訂單"
      parameters:
        - in: path
          name: invitation_code
          schema:
            type: string
          required: true
        - in: path
          name: participant_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: create the item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/participant"
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/responses/BadRequest"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NotFound"    
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Internal Server Error"
    post:
      tags:
        - "發起人/參與人-操作品項於訂單"
      parameters:
        - in: path
          name: invitation_code
          schema:
            type: string
          required: true
        - in: path
          name: participant_id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref:  "#/components/schemas/participant"
      responses:
        '201':
          description: create the item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orders"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NotFound"    
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Internal Server Error"
    put:
      tags:
        - "發起人/參與人-操作品項於訂單"
      parameters:
        - in: path
          name: invitation_code
          schema:
            type: string
          required: true
        - in: path
          name: participant_id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref:  "#/components/schemas/participant"
      responses:
        '200':
          description: update the item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orders"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NotFound" 
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Internal Server Error"

  /orders/{orders_id}:
    get:
      tags:
        - "發起人-取得指定訂單的邀請碼"
      summary: Get the invitation code of the specfic order
      description: ""
      parameters:
        - in: path
          name: orders_id
          schema: 
              type: string
          required: true
      responses:
        '200':
          description: get the invitation code in the specfic order
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/responses/BadRequest"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/responses/NotFound"    
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/responses/Internal Server Error"

components:
  schemas:
    restaurants:
      type: object
      properties:
        _id: 
          type: string
        restaurantName:
          type: string
        restaurantURL:
          type: string
        restaurantMenu:
          type: string
      example:
        restaurantName: "50嵐新竹光復店"
        restaurantURL: "restaurant url..."
        restauranMenu: ""
      required:
        - restaurantName
        - restaurantURL

    participant-items:
      type: object
      properties: 
        _id:
          type: string
        itemName:
          type: string
        orderedNum:
          type: integer
          default: 0
      required:
        - itemName
        - orderedNum

    participant:
      type: object
      properties:
        _id:
          type: string
        PID:
          type: string
        items:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/participant-items"
      required:
        - PID
        - items

    order-items:
      type: object
      properties:
        _id:
          type: string
        itemName:
          type: string
        itemPrice:
          type: integer
      example:
        itemName: "珍奶全糖全冰"
        itemPrice: 90
      required:
        - itemName
        - itemPrice

    orders:
      type: object
      properties:
        _id: 
          type: string
        ownerID:
          type: string
        invitationCode:
          type: string
        authority:
          type: boolean
          default: false
        orderStatus:
          type: boolean
          default: true
        orderItem:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/order-items"
        closeTimestamp:
          type: string
        restaurantID:
          type: string
          $ref: "#/components/schemas/restaurants"
        participant:
          type: array
          items:
            type: object
            $ref: "#/components/schemas/participant"

      required:
        - ownerID
        - restaurantName
        - restaurantURL

    Error:
      type: object
      properties:
        code:
          type: integer
        msg:
          type: string
      required:
        - code
        - msg

  responses:
    BadRequest:
      description: "Request is invaild"
      content:
        application/json:
          schema:
            $ref:  "#/components/schemas/Error"

    Unauthoried:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    NotFound:
      description: "The specified resource was not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    
    Internal Server Error:
      description: "Internal Server Error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"